requirements.txt
asgiref==3.4.1
bitstring==3.1.9
certifi==2021.10.8
charset-normalizer==2.0.9
click==7.1.2
Django==3.2.10
idna==3.3
image==1.5.33
Naked==0.1.31
numpy==1.21.5
Pillow==8.4.0
pycryptodome==3.12.0
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.26.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.7
